+++ vasuki/__init__.py
@@ -6,3 +6,4 @@ from vasuki.gen.uuid import generate_uuid4
 from vasuki.gen.ulid import generate_ulid
 from vasuki.gen.naga import generate_naga19
 from vasuki.gen.gibberish import generate_gibberish
+from vasuki.format.sixnibblename.sixnibblename import integer_slug