+++ CHANGES.rst
@@ -11,6 +11,7 @@ in progress
 - Add ``--upper`` and ``--lower`` options for output transformation
 - Add Nagamani19 identifiers
 - Add Gibberish words
+- Add six-nibble-name integer slug
 
 
 2019-06-28 0.0.0